requirements.txt
flask
python-nmap
scapy
mac-vendor-lookup